Don't relay empty changes to SharedBuffer
After implementing 2ecf5cb, we started observing `Cannot read property
opId of undefined` exceptions. This started happening because of an odd
behavior in text-buffer that causes no-op changes to still trigger an
`onDidChange` event.

While I think this should be solved in text-buffer, it's also true that
we are going to move away from non-batched events pretty soon (and,
long-term, `onDidChange` should be deprecated as well).

Hence, this commit simply provides a way of ignoring such no-op events,
thus fixing the error we were observing.

cc: @nathansobo @jasonrudolph