Autoscroll to the host's cursor position when host changes active editor
Fixes #13.

When the editor is not attached to the DOM, Atom will simply ignore
(auto)scroll requests. This made the real-time package exhibit the
behavior described in the above issue, since we were initializing
`BufferBinding` and `EditorBinding` before adding the `TextEditor`
instance to the workspace.

With this commit, wiring up the bindings between the buffer/editor and
their shared counterparts happens after attaching the editor to the DOM,
thus fixing the aforementioned issue.