Initialize sharedEditorsByEditor map when sharing a portal
When we close a portal, we invoke dispose() on the SharedEditor objects
in the map, but the objects remain in the map. If the host shares a new
portal, we need to ensure that we don't attempt to reuse the
SharedEditor objects that have already been disposed. We can meet this
goal by creating a new sharedEditorsByEditor map each time we share a
new portal.