Clear tail of remote markers if becoming empty after a text change
Originally, we attempted to solve this by relaying the `exclusive`
parameter from the site owning the marker. This was, however, inaccurate
because selections only become exclusive if text changes and the
selected region becomes empty.

For this reason, similarly to what we do in
[selection.coffee](https://github.com/atom/atom/blob/a3e98d54e3d24bbc5c4a67a135cfae2744805285/src/selection.coffee),
we will observe remote selection markers, and if they move as a result
of a text change, we will clear the tail if their range becomes empty.