Remove "leave portal" command
We want to drive people toward the popover, while also ensuring that you
can perform common operations via a command or keyboard shortcut. In the
case of leaving a portal, the user can click the "leave" button in the
popover *or* they can close the portal tab (e.g., "command+w"). Both
actions will result in the user leaving the portal.