Improve UX when using commands
With this commit we will now:

* Stop showing a notification on `real-time:share-portal`. Instead, the
popover will be revealed and we'll share the portal automatically on
behalf of the user. Note that we will also stop automatically copying
the portal id to the clipboard.
* Stop showing a notification on `real-time:close-portal`. Instead, the
popover will be revealed and we'll stop sharing automatically on behalf
of the user.
* Stop showing portal id prompt on `real-time:join-portal`. Instead, the
popover will be revealed and we'll focus the portal ID text entry on
behalf of the user.