Update test to compare *visible* decorated cursor ranges
We use `getCursorDecoratedRanges` to get the cursors that a portal
participant sees in their editor. We then compare the cursors that one
participant sees to the cursors that another participant sees to ensure
that the participants are seeing the same thing. With the introduction
of tethering, if intentionally hide the cursor for any participant that
is following another participant. The hidden cursors are present in
marker's decorations, but the hidden cursors have an opacity of 0. To
make sure we're only comparing *visible* cursor ranges, we need to teach
getCursorDecoratedRanges to exclude ranges that have a hidden cursor.