Depend on the keytar module bundled with Atom
To avoid the issues that users often encounter (especially on Windows) when installing a package that contains native code, we're choosing to depend on the keytar module that's bundled with Atom. This is admittedly a bit of a hack, but it feels like the pragmatic choice for now given the effort that would be required to set up something like node-pre-gyp to distribute platform-specific builds of keytar.

Closes https://github.com/atom/real-time/issues/111.