Fixes https://github.com/atom/real-time/issues/139
Prior to this change, we would render the HostPortalComponent with a
height that is determined by the height of the toggle switch, and then
the host avatar img would get rendered, and that would increase the
height of HostPortalComponent. By setting the height of the host avatar
img in CSS, the img's height will remain constant when the img element
is first rendered _and_ after the avatar is fetched from the
GitHub CDN and rendered inside the popover. 😅