Use an approximation to determine the visibility of a position
This will allow us to avoid the cost of calling
`pixelPositionForScreenPosition` on the editor element, which causes a
re-render of the entire editor and can be pretty bad in terms of
performance.

The tradeoff here is that we could return wrong values when a
non-monospace font is used, or when a line has many non-monospaced
ligatures. We think this is the right choice, though, considering its
performance benefits.