Don't disconnect tether when resizing editor
When the user scrolls left or right, it's reasonable to assume that the
user wants to navigate around on their own, so we should disconnect the
tether at that point. However, when the user simply resizes the editor,
that probably doesn't imply a desire to navigate on their own; instead,
they're probably just trying to adjust their viewport to their liking.

This change resolves the intermittent test failures described in
https://github.com/atom/real-time/pull/161#issuecomment-340902180