Relay active editor changes only if they're manually initiated by users
This prevents a bug that was previously causing followers to disconnect
their tether when the leader rapidly switched among tabs and destroyed
them before followers had the chance to download the remote
editors/buffers associated with them.

In particular, followers would receive two messages indicating that the
leader wanted to switch to a new tab and that they wanted to close the
previous one. When trying to download the editor associated with the new
tab, the guest would find out that such editor did not exist anymore
and, therefore, would skip that message and only process the second
message indicating that the old editor needed to be removed. In doing
so, it would remove the old editor from the workspace and call
`portal.activateEditorProxy` with the editor proxy associated with the
newly active pane item in the workspace. In turn, this would cause the
follower to disconnect their tether, even if the two peers did not
really want to start browsing tabs on their own.

With this commit we are explicitly preventing this problem from
happening by not relaying active editor changes to `portal` when such
changes happen as a result of an update to the workspace caused by the
leader while the tether is retracted.