🐛 Deremotify Buffer when deremotifying its last remote Editor
Prior to this change, if I host split an editor and then closed one of
the two tabs for the editor, the remote editor on the host would have
its title changed to "untitled".

With this change, we keep track of how many remote editors we have for a
remote buffer. When we deremotify the last editor for the buffer, we
deremotify the buffer as well.