Leave portal when guest closes last remote editor
When the guest closes the last remote editor or the portal's empty pane
item, leave the portal.

Prior to this change, if the guest closes the last remote editor, the
guest would be in limbo: they're still in the portal, but they don't see
any remote editors. Their only remedy was to manually leave and re-join
the portal.

In an upcoming release, we hope to integrate the guest's fuzzy finder
with the portal. Once that functionality is in place:
- If the guest closes the last remote editor, the guest could use the
  fuzzy finder to open any of the available remote editors, thus
  preventing the limbo state described above
- Because of this, we could remove the behavior implemented in this
  commit (i.e., we would no longer have to leave the portal when the
  guest closes the last remote editor)