Replace getRemoteBuffers with getRemoteEditors
We can only open local editors for which a remote editor already exists,
so it makes sense to expose editors instead of remote buffers. In the
future we may want to decouple remote editors from local editors (e.g.
if we want to allow opening arbitrary files that aren't necessarily in
the host's workspace), but we will probably expose a different API for
that when that time comes.