Change strategy to determine if editor was manually destroyed
This fixes a regression introduced in the previous commit that was
causing notification to not appear when host closed the portal or lost
the connection.

Specifically, these notifications were not appearing because
GuestPortalBinding interpreted closing remote tabs as being a
user-initiated action, instead of as a result of the host tearing down
their portal. Thus, when the last tab was being closed the guest would
mistakenly leave the portal instead of being kicked out of it.

With this commit we rely on when `editorBinding.onDidDispose` is fired
to determine if the editor was destroyed manually or by the host. In
particular, if `onDidDispose` is called before actually destroying the
pane item associated with the editor, it means that
`EditorProxy.dispose` was called first and the host wanted to tear down
that particular editor. Otherwise, if we haven't processed the binding
`onDidDispose` event when an editor is destroyed it means the editor was
destroyed by the user.