🔥 Remove obsolete logic for "deremotifying" guest editors
Prior to #319, when the host closed the portal, Teletype would transform
the remote editors in the guest's workspace into local untitled/unsaved
editors. As of #319, when the host closed the portal, Teletype *closes*
remote editors in the guest's workspace. As a result, we no longer need
the logic for transforming the remote editors into local editors.