Simplify relaying of active editor changes
Instead of using `removeEditorProxy` to understand whether an editor was
removed manually or not from a guest portal, we are moving the code that
manages the destruction of an editor from EditorBinding to
GuestPortalBinding.

This allows to understand the order in which the editor and the editor
proxy destruction events occur. If the former occurs before the latter,
it means that the user has manually destroyed the editor. If the events
occur in the opposite order, it means that the host has destroyed the
editor. These events are then used to determine whether the tether
should be detached after the active pane item changes.

This refactoring will pave the way to remove `addEditorProxy` and
`removeEditorProxy`.