Don't relay selections that are destroyed when creation event is emitted
This is a workaround to a bug that exists in core where, if a marker is
destroyed during the handling of a `onDidCreateMarker` event, subsequent
observers are still notified of the creation event, even though the
marker is already destroyed.

This happens frequently when two selections overlap, whereby Atom merges
the selections, destroying the most recent one inside of an
`onDidCreateMarker` event handler. This should really be fixed in core
(i.e. when iterating over observers to notify them of marker creation,
if any observer causes the marker to be destroyed, don't notify the
remaining observers).

In the meantime this feels like a reasonable workaround for Teletype,
and will prevent participants from seeing spurious cursors when another
participant creates selections that end up overlapping with each other.

Co-authored-by: Jason Rudolph <jason@jasonrudolph.com>