Simplify layout for site position avatars
Prior to this change, we rendered three different site position
components (upper right, middle right, and lower right), and a
participant's location in the current file determined which site
position component they appeared in. This seemed like useful behavior
when we were first designing the Teletype UI, but in practice, this
behavior isn't useful enough to justify the complexity.

With the changes in this commit, we now render a single site position
component in the lower right corner of the active text editor, and we
show all site participant avatars in that component.