Display avatar for sites viewing non-portal items
Prior to this commit, if a site was viewing a non-portal pane item
(e.g., the Atom "Settings" tab), the SitePositionsComponent did not show
the site's avatar. If you know that a given person is in the portal, it
can be confusing to not see their avatar in the SitePositionsComponent.

With the changes in this commit, the SitePositionsComponent will now
show the avatars for *all* sites. If a site is viewing a non-portal pane
item, the SitePositionsComponent gives the avatar a lighter opacity to
distinguish it from the avatars of sites that are currently viewing an
editor associated with the portal.

As part of this change, this commit consolidates SitePositionsController
and SitePositionsComponent into one class: SitePositionsComponent.