Always use findPortalId when reading user input
This will add support for both URLs and plain IDs. Eventually we should
probably deprecate using IDs, but until it becomes a maintenance burden
it seems okay to not be strict.

Note that we have only added a simple test to show that the system is
capable of handling portal IDs in addition to URIs. Since this won't be
the "happy path" anymore, it seems okay to focus tests more around URIs.