Split into two RFCs: One for portal URLs & one for the "buddy list"
All of the review discussion in #344 is focused on the behavior of the
"buddy list", and that portion of the RFC needs more refinement before
we move forward. The rest of the existing RFC involves replacing the
existing use of portal IDs with more user-friendly portal URLs. Support
for portal URLs shipped today in Teletype 0.11.

With that in mind, this commit splits the existing RFC into two separate
RFCs:

- RFC-003 covers the introduction of portal URLs, and that RFC has a
status of "Implemented"
- RFC-004 contains the original "buddy list" proposal. That RFC still
has a status of "Proposed". Future commits will update that proposal to
incorporate the review feedback provided in #344.