Improve UX when inviting mutually-trusted collaborator
Update RFC to resolve the following drawback:

> Imagine you code together every day with a particular coworker.
> When you add them as a trusted collaborator, and you invite them to
> join your portal, and they don't join right away, you can't tell if
> it's because they're not online, or if they declined your invitation,
> or if they didn't see your invitation, or if they don't have you
> listed as a trusted collaborator. If you were expecting them to join
> your portal, you have to reach out to them via some other
> communication medium to find out why they haven't joined.

When you opt-in to adding someone as a trusted collaborator, Teletype
will inform you that the person will be able to see whether you're
online and will be able to send you in-Atom portal invitations. Because
you're _opting_ to mark them as a trusted collaborator, Teletype will
also inform them when if you decline an invitation.